86eqddcca #Comments Addressed - Image Management & Resource Management UI
