+++ packages/core-library/system/app/internal/blocks/Hub/Settings/SettingsManagement/steps/ImageManagement/ImageManagement.tsx
@@ -3,17 +3,22 @@
  * Reuse as a whole or in part is prohibited without permission.
  * Created by the Software Strategy & Development Division
  */
-import { Box, Button, Typography } from "@mui/material";
+import { Box, Typography } from "@mui/material";
 import { SettingsSelectionType } from "../../types";
-import { Card } from "../../../../../../../../../components";
+import { Card, Button, DialogBox } from "../../../../../../../../../components";
 import Image from "next/image";
 import { cardItems, LogoData } from "./constant/constant";
-import { DialogBox } from "../../../../../../../../../components/Dialog/DialogBox";
 import { CSSProperties, useState } from "react";
-import { LogoSelectionDialog } from "./components/LogoSelectionDialog";
+import { LogoSelectionDialog } from "./content/LogoSelectionDialog";
 import { StaticImport } from "next/dist/shared/lib/get-img-props";
-import { ImageGallery } from "./components/ImageGallery";
-import { PageSelectionDialog } from "./components/PageSelectionDialog";
+import { ImageGallery } from "./content/ImageGallery";
+import { PageSelectionDialog } from "./content/PageSelectionDialog";
+import {
+  AssetCardDisplayType,
+  categoryToDisplayType,
+  ImageManagementCategoryTypes,
+  LogoSelectProps,
+} from "./types/types";
 
 interface Props {
   nextStep(values: Partial<SettingsSelectionType>): void;
@@ -21,12 +26,6 @@ interface Props {
   values: Partial<SettingsSelectionType>;
   previous: () => void;
   reset: () => void;
-}
-
-interface LogoSelectProps {
-  image: string | StaticImport;
-  bgColor: string;
-  imgStyle: CSSProperties | undefined;
 }
 
 export const ImageManagement: React.FC<Props> = ({
@@ -42,28 +41,6 @@ export const ImageManagement: React.FC<Props> = ({
   const [logoSxContainer, seLogoSxContainer] = useState<
     CSSProperties | undefined
   >();
-
-  const cardBGSx = {
-    bgcolor: "#d9d9d9",
-    borderRadius: "10px",
-    display: "flex",
-    flexDirection: "column",
-    padding: "10px",
-    gap: "10px",
-    flexGrow: 1,
-    flexBasis: "250px",
-    cursor: "pointer",
-  };
-  const cardItemSx = {
-    bgcolor: "#fefefe",
-    borderRadius: "10px",
-    height: "150px",
-    flexGrow: 1,
-    display: "flex",
-    padding: "10px",
-    alignItems: "center",
-    justifyContent: "center",
-  };
 
   const handleLogoSelect = ({ image, bgColor, imgStyle }: LogoSelectProps) => {
     setShowModal(true);
@@ -97,126 +74,31 @@ export const ImageManagement: React.FC<Props> = ({
         <Typography variant="h2" sx={{ fontWeight: "600" }}>
           Content Image Assignation
         </Typography>
-        <Box>
-          <Typography variant="h2" sx={{ fontWeight: "600", marginY: "20px" }}>
-            Logo
-          </Typography>
-          <Box
-            sx={{
-              bgcolor: "#d9d9d9",
-              borderRadius: "10px",
-              display: "flex",
-              padding: "10px",
-              gap: "10px",
-              flexWrap: "wrap",
-            }}
-          >
-            {LogoData.length > 0 &&
-              LogoData.map((card, index) => (
-                <Card
-                  key={index}
-                  onClick={() =>
-                    handleLogoSelect({
-                      image: card.imgSrc,
-                      bgColor: card.bgColor,
-                      imgStyle: card.imgStyle,
-                    })
-                  }
-                  sx={{
-                    borderRadius: "10px",
-                    height: "150px",
-                    flexGrow: 1,
-                    display: "flex",
-                    flexDirection: "column",
-                    flexBasis: 200,
-                    bgcolor: card.bgColor,
-                    cursor: "pointer",
-                  }}
-                >
-                  <Typography
-                    sx={{
-                      fontWeight: "600",
-                      color: card.bgColor === "#404040" ? "white" : "black",
-                    }}
-                  >
-                    {card.title}
-                  </Typography>
-                  <Box
-                    sx={{
-                      display: "flex",
-                      flexGrow: 1,
-                      justifyContent: "center",
-                      alignItems: "center",
-                      marginTop: "20px",
-                    }}
-                  >
-                    <Image
-                      src={card.imgSrc}
-                      alt={card.imgAlt}
-                      style={card.imgStyle || {}}
-                    />
-                  </Box>
-                </Card>
-              ))}
-          </Box>
-          <Box
-            sx={{
-              display: "flex",
-              gap: "10px",
-              marginY: "20px",
-              flexWrap: "wrap",
-            }}
-          >
-            {cardItems.length > 0 &&
-              cardItems.map((item, index) => (
-                <Box
-                  key={index}
-                  sx={cardBGSx}
-                  onClick={() => handleCardDialog(item.content)}
-                >
-                  <Typography sx={{ fontWeight: "600" }}>
-                    {item.title}
-                  </Typography>
-                  <Box sx={cardItemSx}>
-                    <Typography variant="h3" sx={{ fontStyle: "italic" }}>
-                      {item.content}
-                    </Typography>
-                  </Box>
-                </Box>
-              ))}
-          </Box>
-        </Box>
+        <AssetCardDisplay
+          handleLogoSelect={handleLogoSelect}
+          handleCardDialog={handleCardDialog}
+        />
       </Box>
-
-      {categoryToDisplay({
-        category: categoryToChange,
-        logoContainer,
-        bgColorContainer,
-        logoSxContainer,
-        setShowModal,
-        showModal,
-      })}
+      <CategoryDisplay
+        category={categoryToChange as ImageManagementCategoryTypes}
+        logoContainer={logoContainer}
+        bgColorContainer={bgColorContainer}
+        logoSxContainer={logoSxContainer}
+        setShowModal={setShowModal}
+        showModal={showModal}
+      />
     </Box>
   );
 };
 
-type categoryToDisplayType = {
-  category?: string;
-  logoContainer?: string | StaticImport;
-  bgColorContainer?: string;
-  logoSxContainer: CSSProperties | undefined;
-  setShowModal?: (value: boolean) => void;
-  showModal?: boolean;
-};
-
-function categoryToDisplay({
+const CategoryDisplay = ({
   category,
   logoContainer,
   bgColorContainer,
   logoSxContainer,
   setShowModal,
   showModal = false,
-}: categoryToDisplayType) {
+}: categoryToDisplayType) => {
   switch (category) {
     case "Logo":
       return (
@@ -225,17 +107,16 @@ function categoryToDisplay({
           header="Replace Image"
           open={showModal}
           borderRadius="16px"
-          children={
-            <LogoSelectionDialog
-              image={logoContainer}
-              bgColor={bgColorContainer}
-              imgStyle={logoSxContainer}
-            />
-          }
           maxWidth="sm"
           overflowContent="hidden"
           ContentHeight="270px"
-        />
+        >
+          <LogoSelectionDialog
+            image={logoContainer}
+            bgColor={bgColorContainer}
+            imgStyle={logoSxContainer}
+          />
+        </DialogBox>
       );
     case "Images":
       return (
@@ -244,10 +125,11 @@ function categoryToDisplay({
           header="Image Gallery"
           open={showModal}
           borderRadius="16px"
-          children={<ImageGallery isEditable />}
           maxWidth="md"
           ContentHeight="650px"
-        />
+        >
+          <ImageGallery isEditable />
+        </DialogBox>
       );
     case "Web Customer":
       return (
@@ -255,27 +137,130 @@ function categoryToDisplay({
           handleClose={() => setShowModal && setShowModal(false)}
           open={showModal}
           borderRadius="16px"
-          children={<PageSelectionDialog app="Web Customer" />}
           maxWidth="md"
           ContentHeight="650px"
-        />
+        >
+          <PageSelectionDialog selectedCard="Web Customer" />
+        </DialogBox>
       );
-    //-----------------------------------------------------------
-    // This is for preparation if Backoffice will also have an Image
-    //-----------------------------------------------------------
-    // case "Web Backoffice":
-    //   return (
-    //     <DialogBox
-    //       handleClose={() => setShowModal && setShowModal(false)}
-    //       header="Image Gallery"
-    //       open={showModal}
-    //       borderRadius="16px"
-    //       children={<PageSelectionDialog app="Web Backoffice" />}
-    //       maxWidth="md"
-    //       ContentHeight="650px"
-    //     />
-    //   );
     default:
       return null;
   }
-}
+};
+
+const AssetCardDisplay = ({
+  handleLogoSelect,
+  handleCardDialog,
+}: AssetCardDisplayType) => {
+  const cardBGSx = {
+    bgcolor: "#d9d9d9",
+    borderRadius: "10px",
+    display: "flex",
+    flexDirection: "column",
+    padding: "10px",
+    gap: "10px",
+    flexGrow: 1,
+    flexBasis: "250px",
+    cursor: "pointer",
+  };
+
+  const cardItemSx = {
+    bgcolor: "#fefefe",
+    borderRadius: "10px",
+    height: "150px",
+    flexGrow: 1,
+    display: "flex",
+    padding: "10px",
+    alignItems: "center",
+    justifyContent: "center",
+  };
+  return (
+    <Box>
+      <Typography variant="h2" sx={{ fontWeight: "600", marginY: "20px" }}>
+        Logo
+      </Typography>
+      <Box
+        sx={{
+          bgcolor: "#d9d9d9",
+          borderRadius: "10px",
+          display: "flex",
+          padding: "10px",
+          gap: "10px",
+          flexWrap: "wrap",
+        }}
+      >
+        {LogoData.length > 0 &&
+          LogoData.map((card, index) => (
+            <Card
+              key={index}
+              onClick={() =>
+                handleLogoSelect({
+                  image: card.imgSrc,
+                  bgColor: card.bgColor,
+                  imgStyle: card.imgStyle,
+                })
+              }
+              sx={{
+                borderRadius: "10px",
+                height: "150px",
+                flexGrow: 1,
+                display: "flex",
+                flexDirection: "column",
+                flexBasis: 200,
+                bgcolor: card.bgColor,
+                cursor: "pointer",
+              }}
+            >
+              <Typography
+                sx={{
+                  fontWeight: "600",
+                  color: card.bgColor === "#404040" ? "white" : "black",
+                }}
+              >
+                {card.title}
+              </Typography>
+              <Box
+                sx={{
+                  display: "flex",
+                  flexGrow: 1,
+                  justifyContent: "center",
+                  alignItems: "center",
+                  marginTop: "20px",
+                }}
+              >
+                <Image
+                  src={card.imgSrc}
+                  alt={card.imgAlt}
+                  style={card.imgStyle || {}}
+                />
+              </Box>
+            </Card>
+          ))}
+      </Box>
+      <Box
+        sx={{
+          display: "flex",
+          gap: "10px",
+          marginY: "20px",
+          flexWrap: "wrap",
+        }}
+      >
+        {cardItems.length > 0 &&
+          cardItems.map((item, index) => (
+            <Box
+              key={index}
+              sx={cardBGSx}
+              onClick={() => handleCardDialog(item.content)}
+            >
+              <Typography sx={{ fontWeight: "600" }}>{item.title}</Typography>
+              <Box sx={cardItemSx}>
+                <Typography variant="h3" sx={{ fontStyle: "italic" }}>
+                  {item.content}
+                </Typography>
+              </Box>
+            </Box>
+          ))}
+      </Box>
+    </Box>
+  );
+};
